Address PR review feedback
Copilot:
- Move formatter DTO polymorphism tests out of the oversized client-package
  PolymorphicSerializationTest into a dedicated
  com.otilm.api.model.connector.signatures.formatter.FormatterPolymorphicSerializationTest,
  matching the repo convention for connector DTO tests.
- Drop the now-unused connector.signatures.formatter imports from
  PolymorphicSerializationTest.
